+++ griptape/loaders/base_text_loader.py
@@ -7,7 +7,7 @@ from attrs import define, field, Factory
 from pathlib import Path
 
 from griptape.artifacts import TextArtifact
-from griptape.chunkers import TextChunker
+from griptape.chunkers import TextChunker, BaseChunker
 from griptape.drivers import BaseEmbeddingDriver
 from griptape.loaders import BaseLoader
 from griptape.tokenizers import OpenAiTokenizer
@@ -24,7 +24,7 @@ class BaseTextLoader(BaseLoader, ABC):
         default=Factory(lambda self: round(self.tokenizer.max_tokens * self.MAX_TOKEN_RATIO), takes_self=True),
         kw_only=True,
     )
-    chunker: TextChunker = field(
+    chunker: BaseChunker = field(
         default=Factory(
             lambda self: TextChunker(tokenizer=self.tokenizer, max_tokens=self.max_tokens), takes_self=True
         ),