+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.23.2] - 2024-03-14
+
+### Fixed
+- Deprecation warnings not displaying for `Structure.prompt_driver`, `Structure.embedding_driver`, and `Structure.stream`.
+- `DummyException` error message not fully displaying.
+
 ## [0.23.1] - 2024-03-07
 
 ### Fixed